Prove finite adeles have the module topology (#633)
Defines a Module instance for restricted products of modules over a
restricted product of rings. Show that this has the module topology when
it's finitely generated and all of the components have the module
topology. Use this to show that `A^f L` has the `A^f K`-module topology.