GL_2 over finite adeles is restricted product of local GL_2(K_v)s (#659)
This is [Theorem
12.20](https://imperialcollegelondon.github.io/FLT/blueprint/HeckeOperator_project.html#nolean-GL2-finite-adeles)
on the decomposition of global GL_2 as a restricted product of local
GL_2. It was needed (not strictly necessary, but would make sense to
have first) in #584. It also includes definition and statements of
[Lemma
12.16](https://imperialcollegelondon.github.io/FLT/blueprint/HeckeOperator_project.html#nolean-compactopen-matrix)
on M_2(O_v) as a (compact open) subring of M_2(K_v).

---------

Co-authored-by: Kevin Buzzard <k.buzzard@imperial.ac.uk>