Explicit local double coset decomposition in preparation for Hecke operators (#660)
This is [Lemma
12.19](https://imperialcollegelondon.github.io/FLT/blueprint/HeckeOperator_project.html#nolean-U1-coset-decomposition)
of the blueprint, which is the essential component of #584. It is in a
separate file due to its length, but also because it is a purely local
computation in preparation for the commutativity of Hecke operators, and
there didn't seem to be anywhere else suitable to place it.