chore: tidy up Galois representation on torsion points (#750)
This PR moves away from the `IsSimpleModule` approach to Mazur's and
Wiles' theorems, and instead sticks to `GaloisRep.IsIrreducible`. It
also tidies up a few sorries, pulls out Frey packages into their own
file to sort out an import issue, fills in the proof of Wiles' theorem
on reducibility of p-torsion of the Frey curve modulo many other hard
theorems, fixes a howler in the definition of the p-torsion of the Frey
curve (we were considering points of Q not Qbar), defines the Galois
action on the Qbar points and fills in a few more sorries.