start on fundamental domain (#797)
We need to prove that if U is a sufficiently large subset of a locally
compact topological abelian group X then U doesn't inject into X/Gamma
where Gamma is a closed subgroup with compact quotient.

There are two ways to proceed. The first goes via Blichfeldt's theorem
as stated in mathlib
[MeasureTheory.exists_pair_mem_lattice_not_disjoint_vadd](https://leanprover-community.github.io/mathlib4_docs/Mathlib/MeasureTheory/Group/GeometryOfNumbers.html#MeasureTheory.exists_pair_mem_lattice_not_disjoint_vadd)
and this needs no extra assumptions on X but it does need a fundamental
domain; this PR establishes a fundamental domain for $\mathbb{Q}$ in
$\mathbb{A}$.

The second is via a more general result in measure theory, which needs
that X is a Polish space and in particular second countable; in this PR
we establish that the adeles of a number field are second countable.